pyproject.toml
[tool.poetry]
name = "airflow-test"
version = "0.1.0"
description = ""
authors = ["강재호 <jhkang@illunex.com>"]
readme = "README.md"
packages = [{include = "airflow_test"}]

[tool.poetry.dependencies]
python = "3.8.17"
apache-airflow = {extras = ["celery"], version = "^2.7.0"}
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
black = "^23.7.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
